pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regime-forecast-lite"
version = "0.1.0"
description = "Streaming one-step-ahead forecaster with a leakage-safe pipeline and optional FastAPI service."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

authors = [
  { name = "j123and" }
]

keywords = ["time-series", "forecasting", "change-point", "conformal", "online-learning", "fastapi"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers"
]

# ——— Core kept thin: safe for library users ———
dependencies = [
  "numpy>=1.26,<3.0",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
# FastAPI service + metrics + client for tests
service = [
  "fastapi>=0.110,<0.120",        # narrow to a minor series
  "pydantic>=2.6,<3",             # explicit pin to match FastAPI v2
  "uvicorn[standard]>=0.30,<1.0", # better defaults
  "prometheus-client>=0.20,<1.0",
  "httpx>=0.24,<1.0",
]

# Plotting (only needed for plots)
plot = ["matplotlib>=3.8,<4.0"]

# Market data fetcher
market = ["yfinance>=0.2.40,<0.3"]

# Backtest convenience (CSV works without these; Parquet needs these)
backtest = [
  "pandas>=2.2,<3.0",
  "pyarrow>=14,<19",
]

# Dev tools ONLY (no self-references here)
dev = [
  "pytest>=8,<9",
  "ruff>=0.5,<1.0",
  "mypy>=1.10,<2.0",
  "tqdm>=4.66,<5.0",
  "types-PyYAML>=6.0.12,<7.0",
  "pre-commit>=3.7,<4.0"
]

[project.urls]
Homepage = "https://github.com/j123and/regime-forecast-lite"
Repository = "https://github.com/j123and/regime-forecast-lite"

# CLI entry points (so users don’t need `python -m`)
[project.scripts]
regime-backtest = "backtest.cli:main"
regime-sim = "data.sim_cp:main"           # REMOVE if that module isn't shipping
regime-fetch-yahoo = "data.yahoo_fetch:main"  # REMOVE if that module isn't shipping

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "scripts*", "examples*", "artifacts*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]
